Share LookInside server environment config
