+++ Example/Tuist/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "cfeb7c8d9b98cf02c5164682b7147b6a863867e3bc31c6852b7bb30d46df4994",
+  "originHash" : "7ca1a1e556159cedef2e6b4648b27086aa27dbdb32eb4de468b034ec81c5b69f",
   "pins" : [
     {
       "identity" : "equatable",
@@ -8,15 +8,6 @@
       "state" : {
         "branch" : "main",
         "revision" : "8f1dccead65d97173ed695c6c1ba9d0570d35619"
-      }
-    },
-    {
-      "identity" : "lookinside-release",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/LookInsideApp/LookInside-Release.git",
-      "state" : {
-        "revision" : "7514cfa8bd24a78f74d8b982c429e7d61ef58134",
-        "version" : "0.2.2"
       }
     },
     {
